Redecorate apps for all executors, to get task label in monitoring
Prior to this PR, task names appeared in the monitoring database when uisng
work queue executor because a run_command-like app was created for every
invocation, and that app was customised to have the bps task label attached.

With other executors, that code path was not used, and so other executors
default to using the parsl default of name of the underlying Python function
which in the gen3_workflow case was `run_command`.

This PR switches to making a new app for every task invocation for all
executor types, and applies the same task label modification. This makes the
task label available in monitoring with all executor types.